requirements.txt
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cloudscraper==1.2.48
colorama==0.4.4
Flask==1.1.2
gallery-dl==1.16.0
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
LEPL==5.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
pylint==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
